FIX do not mysteriously close void elements
Upon outputting a void element, the html5 library in use previous to
this commit would mysteriously append closing tags to void elements -
i.e. <img /> would become <img></img>.

Reading upon the internet at large, this seems to be a known issue with
PHP's DOMDocument library, however I could not find a reason why the
html5 library at it's standard would not recreate the issue, however it
remained would within our framework (requiring only silverstripe/recipe-core
as a baseline to test this).

So this commit has found and employed the use of another library with
the same goal - making html5 easy to process. It appears to succeed
where the previous did not - the test previously added to trigger the
described undesirable functionality is now passing - there are no longer
any invalid closing tags present in HTMLValue output.

Please note that while the version of the new html5 processing library
in use is in fact an out of date one, it is the last version compatible
with PHP 5.6 - which we still support at this time.