revert back to masterminds/html5 to debug issues
which appear to be environmental. This dependency was previously removed
because of the issues caused by it's parsing, however the replacement
dependency also had some improper output which prevents us from using it
(and furthermore was an out of support version). So we need to debug this,
in which case I need to commit so I can share with fellow team members the
issue at hand.